Move tests to their own class
